api/requirements.txt
fastapi>=0.63.0,<0.64.0
gino>=1.0.0,<1.1.0
PyNaCl>=1.4.0,<1.5.0
uvicorn[standard]>=0.13.0,<0.14.0
